requirements.txt
Django==4.2.6
environs==9.5.0
Pillow==10.1.0
django-admin-sortable2==2.1.10

